chore: split tests from api compare
